pyproject.toml
[tool.poetry]
name = "solana_indexer"
version = "0.1.0"
description = ""
authors = ["Landon Gingerich <landon.ging@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
solana = "^0.34.3"
colorlog = "^6.8.2"
solders = "^0.21.0"
polars = "^1.6.0"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
